extract script-related types and funcs from main.go
